+++ lib/WebServer/index.js
@@ -1,15 +1,35 @@
-const http = require("./http.js");
+const async = require("async");
+const web = require("./web.js");
+const patchServer = require("./patchServer.js");
 const logger = require("../../src/logger.js");
 
 function start(callback) {
   /* Start the NPS servers */
-  http.start(err => {
-    if (err) {
-      throw err;
+
+  async.series(
+    {
+      patchServer: function(callback) {
+        patchServer.start(err => {
+          if (err) {
+            throw err;
+          }
+          callback(null);
+        });
+      },
+      web: function(callback) {
+        web.start(err => {
+          if (err) {
+            throw err;
+          }
+          callback(null);
+        });
+      },
+    },
+    function(err, results) {
+      logger.info("AuthLogin Server started");
+      callback();
     }
-    logger.info("HTTP Servers started");
-    callback(null);
-  });
+  );
 }
 
 module.exports = { start };
