+++ lib/WebServer/web.js
@@ -7,7 +7,6 @@ const fs = require("fs");
 const configurationFile = require("../../config/config.json");
 
 const logger = require("../../src/logger.js");
-const patchServer = require("./patch_server.js");
 
 function generateShardList(config) {
   return `[The Clocktower]
@@ -54,24 +53,6 @@ function start(callback) {
   app.use(bodyParser.json()); // support json encoded bodies
   app.use(bodyParser.urlencoded({ extended: true })); // support encoded bodies
 
-  app.post("/games/EA_Seattle/MotorCity/UpdateInfo", (req, res) => {
-    const response = patchServer.patchUpdateInfo(req);
-    res.set(response.headers);
-    res.send(response.body);
-  });
-
-  app.post("/games/EA_Seattle/MotorCity/NPS", (req, res) => {
-    const response = patchServer.patchNPS(req);
-    res.set(response.headers);
-    res.send(response.body);
-  });
-
-  app.post("/games/EA_Seattle/MotorCity/MCO", (req, res) => {
-    const response = patchServer.patchMCO(req);
-    res.set(response.headers);
-    res.send(response.body);
-  });
-
   app.get("/AuthLogin", (req, res) => {
     res.set("Content-Type", "text/plain");
     res.send("Valid=TRUE\nTicket=d316cd2dd6bf870893dfbaaf17f965884e");
@@ -82,11 +63,6 @@ function start(callback) {
     logger.debug(`Method: ${req.method}`);
     logger.debug(`Url: ${req.url}`);
     res.send("404");
-  });
-
-  const serverPatch = http.createServer(app);
-  serverPatch.listen("80", "0.0.0.0", () => {
-    logger.info(`Patch server listening on port ${app.get("port")}`);
   });
 
   const httpsServer = https
